requirements.txt
Flask
ply
google-generativeai
python-dotenv
